+++ examples/uploads/app.py
@@ -1,6 +1,7 @@
 import os
 
-from molten import App, RequestData, Route, UploadedFile
+from molten import App, RequestData, Route, UploadedFile, schema
+from molten.openapi import Metadata, OpenAPIHandler, OpenAPIUIHandler
 
 
 def path_to(*segments):
@@ -17,7 +18,29 @@ def upload_one(f: UploadedFile) -> None:
     f.save(path_to("uploads", f.filename))
 
 
+@schema
+class Data:
+    f: UploadedFile
+
+
+def upload_one_schema(data: Data) -> None:
+    data.f.save(path_to("uploads", data.f.filename))
+
+
+get_schema = OpenAPIHandler(
+    metadata=Metadata(
+        title="uploads",
+        description="file upload api",
+        version="0.0.0",
+    ),
+)
+
+get_docs = OpenAPIUIHandler()
+
 app = App(routes=[
     Route("/upload", upload, method="POST"),
     Route("/upload-one", upload_one, method="POST"),
+    Route("/upload-one-schema", upload_one_schema, method="POST"),
+    Route("/_schema", get_schema),
+    Route("/_docs", get_docs),
 ])