+++ docs/source/changelog.rst
@@ -19,6 +19,14 @@ Added
 * The test client now supports file uploads via the ``file`` keyword
   argument.
 
+Fixed
+^^^^^
+
+* |UploadedFile| now renders a file upload element in OpenAPI
+documents.  (`#23`_)
+
+.. _#23: https://github.com/Bogdanp/molten/issues/23
+
 `0.7.1`_ -- 2018-10-15
 ----------------------